ecomhub/requirements.txt
admin_tools==0.1
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.2
datetime-truncate==1.1.1
Django==5.2
django-admin-charts==1.5.0
django-admin-tools==0.9.3
django-bower==5.2.0
django-filter==25.1
django-jazzmin==3.0.1
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jwcrypto==1.5.6
Markdown==3.7
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
